Refactor creation controller into extensions
